face_server/requirements.txt
Flask>=2.0.0
Flask-Cors>=3.0.10
python-multipart>=0.0.5
pymongo>=4.0.1
face-recognition>=1.3.0
python-dotenv>=0.19.0
numpy>=1.19.0
